pyproject.toml
[project]
name = "hacknation-2025-skillsense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "plotly>=6.4.0",
    "pygithub>=2.8.1",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "spacy>=3.8.8",
    "streamlit>=1.51.0",
    "transformers>=4.57.1",
]
